// admin/types/type.d.ts
export interface Category {
  category_id: string;
  name: string;
  description: string;
  image_url: string;
  created_at: string;
}